% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_nord.R
\name{scale_color_nord}
\alias{scale_color_nord}
\title{Nord scale color for ggplot2}
\usage{
scale_color_nord(
  theme = "nord",
  color_key = list(nord = c("#81a1c1", "#a3be8c", "#d08770", "#b48ead", "#ebcb8b",
    "#8fbcbb", "#bf616a", "#b48ead", "#88c0d0", "#d08770", "#81a1c1", "#ebcb8b",
    "#5e81ac", "#a3be8c", "#d5d6ea", "#d7ecd9", "#f5d5cb", "#f6ecf5", "#f3ddf2",
    "#9d465e", "#6f365c", "#75b390", "#ecd6ba", "#e8a9a5", "#697da2"))
)
}
\description{
Nord color
}
\examples{
scale_color_nord(theme="nord")
}
\author{
Sebastian Rauschert
}
