% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nord.R
\name{theme_nord}
\alias{theme_nord}
\title{Theme based on the Nord color scheme for ggplot2}
\usage{
theme_nord(theme = "nord")
}
\description{
Nord theme for ggplot
}
\examples{
theme_nord(theme="dark")
}
\author{
Sebastian Rauschert
}
